refactor(web): unify song cards and extract shared ui primitives
